requirements.txt
gym>=0.17
numpy==1.20
tqdm==4.60.0
pyyaml==3.13
torch==1.8.0
cloudpickle==1.6.0
pandas==1.1.5
matplotlib==3.3.3
seaborn==0.11.1
stable_baselines3==1.0
tslearn==0.5.1.0
